secuXchange/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
datasets==3.5.1
dill==0.3.8
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
grad-cam==1.5.5
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib~=3.9.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy~=1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow~=10.4.0
propcache==0.3.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==80.0.1
six==1.17.0
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm~=4.66.5
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.121
ultralytics-thop==2.0.14
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0

scikit-learn~=1.5.1